LayoutTests/js/dfg-check-array-non-array-expected.txt

Tests that CheckArray(NonArray) actually proves that the input isn't an array.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS baz(["blah"], true) is "blahblah"
PASS successfullyParsed is true

TEST COMPLETE
